Added negatives indexes to.

+++ sprint2/problems/time_control/solution/src/model/roadmap.h
@@ -27,8 +27,8 @@ public:
                             const Velocity& old_velocity);
 private:
     struct MatrixMapCoord {
-        size_t x;
-        size_t y;
+        int64_t x;
+        int64_t y;
 
         bool operator == (const MatrixMapCoord& other) const {
             return (x == other.x) && (y == other.y);
